Renamed toast-types.ts to utility-types.ts & Sync Configs
