+++ src/1-Profile/Signup.tsx
@@ -18,7 +18,7 @@ import { FormInput } from '../Widgets/FormInput/FormInput';
 import { AppStackParamList, ROUTE_NAMES } from '../TypesAndInterfaces/routes';
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
 import Ionicons from 'react-native-vector-icons/Ionicons';
-import { ServerErrorResponse } from '../TypesAndInterfaces/config-sync/api-type-sync/toast-types';
+import { ServerErrorResponse } from '../TypesAndInterfaces/config-sync/api-type-sync/utility-types';
 import ToastQueueManager from '../utilities/ToastQueueManager';
 import { ProfileResponse } from '../TypesAndInterfaces/config-sync/api-type-sync/profile-types';
