+++ src/3-Prayer-Request/PrayerRequestCommentCreate.tsx
@@ -12,7 +12,7 @@ import { DOMAIN } from '@env';
 import axios, { AxiosError, AxiosResponse } from 'axios';
 import { RootState } from '../redux-store';
 import { useAppSelector } from '../TypesAndInterfaces/hooks';
-import { ServerErrorResponse } from '../TypesAndInterfaces/config-sync/api-type-sync/toast-types';
+import { ServerErrorResponse } from '../TypesAndInterfaces/config-sync/api-type-sync/utility-types';
 import ToastQueueManager from '../utilities/ToastQueueManager';
 import { RootSiblingParent } from 'react-native-root-siblings';
