+++ src/TypesAndInterfaces/config-sync/input-config-sync/content-field-config.ts
@@ -68,7 +68,7 @@ export enum ContentSearchRefineEnum {
 * UTILITY METHODS *
 *******************/
 export const extractYouTubeVideoId = (url:string):string|undefined => {
-    const match = url?.match(/(?:https?:\/\/)?(?:www\.)?(?:youtube\.com\/(?:[^\/\n\s]+\/\S+\/|(?:v|e(?:mbed)?)\/|\S*?[?&]v=)|youtu\.be\/)([a-zA-Z0-9_-]{11})/);
+    const match = url.match(/(?:https?:\/\/)?(?:www\.)?(?:youtube\.com\/(?:[^\/\n\s]+\/\S+\/|(?:v|e(?:mbed)?)\/|\S*?[?&]v=)|youtu\.be\/)([a-zA-Z0-9_-]{11})/);
     return match ? match[1] : undefined;
   };
